+++ src/main/java/fr/estecka/variantscit/reload/ModuleDefinition.java
@@ -1,4 +1,4 @@
-package fr.estecka.variantscit;
+package fr.estecka.variantscit.reload;
 
 import java.util.HashMap;
 import java.util.List;
@@ -11,6 +11,7 @@ import com.mojang.serialization.Codec;
 import com.mojang.serialization.DataResult;
 import com.mojang.serialization.MapCodec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
+import fr.estecka.variantscit.VariantsCitMod;
 
 public record ModuleDefinition(
 	Identifier type,
